refactor(users): changed file names and directories
