+++ frontend/src/app/profile/components/user/user.component.scss
